moved searchTagNameInInput and cleanSearchInput functions to utils
