+++ services/utils/classNames.js
@@ -1,0 +1,4 @@
+
+export function classNames(...classes) {
+    return classes.filter(Boolean).join(" ");
+}
